requirements.txt
# Core ML frameworks
torch>=2.0.0
transformers>=4.35.0
datasets>=2.14.0

# Vietnamese NLP preprocessing
vncorenlp>=1.0.3
pyvi>=0.1.1

# Training & Tracking
accelerate>=0.24.0
wandb>=0.16.0

# Data handling
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0

# API & Serving
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.4.0
python-multipart>=0.0.6

# Utilities
pyyaml>=6.0
python-dotenv>=1.0.0
tqdm>=4.66.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Development
pytest>=7.4.0
jupyter>=1.0.0
